Added exception handling and user token verification API

+++ public/admin.js
@@ -28,7 +28,15 @@ document
     }
   });
 
-async function deletePost(id) {
+async function deletePost(id, event) {
+  event.stopPropagation();
+  
+  // Confirm if admin wants to delete this for sure.
+  const confirmation = confirm("Are you sure you want to delete this blog post?");
+  if (!confirmation) {
+    return; 
+  }
+
   const res = await fetch(`/api/blogs/${id}`, {
     method: "DELETE",
     headers: { Authorization: "Bearer " + localStorage.getItem("token") },
@@ -45,17 +53,17 @@ function logout() {
   window.location.href = "login.html";
 }
 
-fetchBlogs();
-
 // Load blogs
 function loadBlogs(blogs) {
   const container = document.getElementById("blogs-container");
+  container.innerHTML = "";
   blogs.forEach((blog) => {
     const blogCard = document.createElement("div");
     blogCard.classList.add("blog-card");
     blogCard.innerHTML = `
 <h3>${blog.title}</h3>
 <p>${truncateContent(blog.content, 200)}</p> 
+<button class="delete-blog" onclick="deletePost(${blog.id}, event)">Delete</button>
 `;
     blogCard.addEventListener("click", () => openPopup(blog));
     container.appendChild(blogCard);
@@ -81,3 +89,42 @@ function openPopup(blog) {
 document.getElementById("close-popup").addEventListener("click", () => {
   document.getElementById("popup").style.display = "none";
 });
+
+document.addEventListener("DOMContentLoaded", async () => {
+  await verifyUserUsingToken(); // Check if user is admin, when page loads
+});
+
+async function verifyUserUsingToken() {
+  const token = localStorage.getItem("token");
+
+  if (!token) {
+    alert("Unauthorized access. Please login first.");
+    window.location.href = "login";
+    return;
+  }
+
+  try {
+    const res = await fetch("/api/auth/verify", {
+      headers: {
+        Authorization: "Bearer " + token,
+      },
+    });
+
+    if (!res.ok) {
+      throw new Error("Invalid token");
+    }
+
+    const data = await res.json();
+
+    if (data.role !== "admin") {
+      alert("Access denied. Admins only!");
+      window.location.href = "blogs"; // Redirect to blogs page
+    }
+    await fetchBlogs();
+  } catch (error) {
+    console.error(error);
+    alert("Session expired or invalid. Please login again.");
+    localStorage.removeItem("token");
+    window.location.href = "login";
+  }
+}
